Install script now installs all existing dotfiles

+++ install.sh
@@ -1,17 +1,29 @@
-#!/usr/bin/env sh
+#!/usr/bin/env bash
 
-dir=`pwd`
+dir=$(pwd)
+files=$(ls -ad .*)
+echo "installing: $files"
 
 cd ~
-for dotfile in .gitconfig .gitexcludes .vimrc .Xdefaults; do
-	if [ -s $dotfile ]
+for dotfile in $files; do
+	echo "test: $dotfile"
+	if [ -f $dir/$dotfile ]
 	then
-		cp -v $dotfile $dotfile.orig
-		rm $dotfile
+		if [ -h $dotfile ]
+		then
+			echo "Remove existing link: $dotfile"
+			rm $dotfile
+		fi
+
+		if [ -s $dotfile ]
+		then
+			cp -v $dotfile $dotfile.orig
+			rm $dotfile
+		fi
+		
+		echo "install: $dotfile"
+		ln -s $dir/$dotfile $dotfile
 	fi
-	
-	echo "install: $dotfile"
-	ln -s $dir/$dotfile $dotfile
 done
 
 cd $dir
